requirements.txt
absl-py==2.1.0
aiosignal==1.3.1
annotated-types==0.7.0
anthropic==0.26.1
anyio==4.4.0
astunparse==1.6.3
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cssutils==2.11.0
cycler==0.12.1
distro==1.9.0
exceptiongroup==1.2.1
filelock==3.14.0
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.5.0
gast==0.5.4
google-pasta==0.2.0
graphviz==0.20.3
grpcio==1.64.0
h11==0.14.0
h5py==3.11.0
html5lib==1.1
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.2
idna==3.7
imageio==2.34.2
Jinja2==3.1.4
jiter==0.4.0
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
keras==3.3.3
kiwisolver==1.4.5
langdetect==1.0.9
lazy_loader==0.4
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
msgpack==1.0.8
namex==0.0.8
networkx==3.3
nltk==3.8.1
numpy==1.26.4
ollama==0.2.1
openai==1.30.4
opencv-python==4.10.0.84
opt-einsum==3.3.0
optree==0.11.0
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.2
pillow==10.3.0
protobuf==4.25.3
psutil==5.9.8
py-cpuinfo==9.0.0
pyarrow==16.1.0
pydantic==2.7.2
pydantic_core==2.18.3
Pygments==2.18.0
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
ray==2.23.0
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
rpds-py==0.18.1
scikit-image==0.24.0
scipy==1.14.0
scrapingbee==2.0.1
seaborn==0.13.2
selenium==4.21.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sympy==1.12.1
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.37.0
termcolor==2.4.0
tifffile==2024.7.2
tokenizers==0.19.1
torch==2.3.1
torchvision==0.18.1
torchviz==0.0.2
tqdm==4.66.4
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.0
tzdata==2024.1
ultralytics==8.2.56
ultralytics-thop==2.0.0
urllib3==2.2.1
webencodings==0.5.1
Werkzeug==3.0.3
wrapt==1.16.0
wsproto==1.2.0
